requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.2
dj-database-url==0.5.0
Django==3.2.5
django-appconf==1.0.4
django-compressor==2.4.1
django-heroku==0.3.1
django-sass-processor==1.0.1
environ==1.0
idna==3.2
libsass==0.21.0
psycopg2==2.7.7
python-dateutil==2.8.2
pytz==2021.1
rcssmin==1.0.6
requests==2.26.0
rjsmin==1.1.0
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
urllib3==1.26.6
whitenoise==5.2.0
